Revert "[tests] Duplicate Microsoft.Diagnostics.NETCore.Client dependent tests for Linux arm"
This reverts commit cb2cacd93bb61b74b36c2d2d26cccec38b51b8f7.
