--- src/tests/JIT/Directed/debugging/debuginfo/tester_legacy.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <DebugType>PdbOnly</DebugType>
-    <Optimize>True</Optimize>
-    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
-    <UnloadabilityIncompatible>true</UnloadabilityIncompatible>
-    <GCStressIncompatible>true</GCStressIncompatible>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="tests_d.ilproj" Aliases="tests_d" />
-    <ProjectReference Include="tests_r.ilproj" Aliases="tests_r" />
-    <ProjectReference Include="attribute.csproj" />
-    <ProjectReference Include="../../../../tracing/eventpipe/common_legacy/common_legacy.csproj" />
-    <Compile Include="$(MSBuildProjectName).cs" />
-  </ItemGroup>
-</Project>
