+++ src/tests/tracing/eventpipe/bigevent/bigevent.cs
@@ -11,7 +11,11 @@ using System.Collections.Generic;
 using Microsoft.Diagnostics.Tracing;
 using Tracing.Tests.Common;
 using Microsoft.Diagnostics.Tracing.Parsers.Clr;
-using Microsoft.Diagnostics.NETCore.Client;
+#if (UPDATED_NETCORE_CLIENT == true)
+    using Microsoft.Diagnostics.NETCore.Client;
+#else
+    using Microsoft.Diagnostics.Tools.RuntimeClient;
+#endif
 
 namespace Tracing.Tests.BigEventsValidation
 {
@@ -45,12 +49,22 @@ namespace Tracing.Tests.BigEventsValidation
         {
             // This test tries to send a big event (>100KB) and checks that the app does not crash
             // See https://github.com/dotnet/runtime/issues/50515 for the regression issue
+#if (UPDATED_NETCORE_CLIENT == true)
             var providers = new List<EventPipeProvider>()
             {
                 new EventPipeProvider("BigEventSource", EventLevel.Verbose)
             };
 
             return IpcTraceTest.RunAndValidateEventCounts(_expectedEventCounts, _eventGeneratingAction, providers, 1024, _Verify);
+#else
+            var providers = new List<Provider>()
+            {
+                new Provider("BigEventSource")
+            };
+
+            var configuration = new SessionConfiguration(circularBufferSizeMB: 1024, format: EventPipeSerializationFormat.NetTrace,  providers: providers);
+            return IpcTraceTest.RunAndValidateEventCounts(_expectedEventCounts, _eventGeneratingAction, configuration, _Verify);
+#endif
         }
 
         private static Dictionary<string, ExpectedEventCount> _expectedEventCounts = new Dictionary<string, ExpectedEventCount>()
