--- src/tests/tracing/eventpipe/bigevent_legacy/bigevent_legacy.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETCoreApp</TargetFrameworkIdentifier>
-    <OutputType>exe</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <JitOptimizationSensitive>true</JitOptimizationSensitive>
-    <UnloadabilityIncompatible>true</UnloadabilityIncompatible>
-    <!-- GCStress timeout: https://github.com/dotnet/runtime/issues/51133 -->
-    <GCStressIncompatible Condition="'$(TargetArchitecture)' == 'arm64' and '$(TargetOS)' == 'OSX'">true</GCStressIncompatible>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildProjectName).cs" />
-    <ProjectReference Include="../common_legacy/common_legacy.csproj" />
-  </ItemGroup>
-</Project>
