--- src/tests/tracing/eventpipe/buffersize_legacy/buffersize_legacy.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETCoreApp</TargetFrameworkIdentifier>
-    <OutputType>exe</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <UnloadabilityIncompatible>true</UnloadabilityIncompatible>
-    <JitOptimizationSensitive>true</JitOptimizationSensitive>
-    <GCStressIncompatible>true</GCStressIncompatible>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildProjectName).cs" />
-    <ProjectReference Include="../common_legacy/common_legacy.csproj" />
-  </ItemGroup>
-</Project>
