+++ src/tests/tracing/eventpipe/common/IpcTraceTest.legacy.cs
