+++ src/tests/tracing/eventpipe/common/common.csproj
@@ -13,10 +13,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="IpcTraceTest.cs" />
+    <Compile Condition="'$(UseUpdatedNETCoreClient)' == 'true'" Include="IpcTraceTest.cs" />
+    <Compile Condition="'$(UseUpdatedNETCoreClient)' != 'true'" Include="IpcTraceTest.legacy.cs" />
     <Compile Include="IpcUtils.cs" />
     <Compile Include="Reverse.cs" />
     <Compile Include="StreamProxy.cs" />
-    <ProjectReference Include="Microsoft.Diagnostics.NETCore.Client/Microsoft.Diagnostics.NETCore.Client.csproj" />
+    <ProjectReference Condition="'$(UseUpdatedNETCoreClient)' == 'true'" Include="Microsoft.Diagnostics.NETCore.Client/Microsoft.Diagnostics.NETCore.Client.csproj" />
   </ItemGroup>
 </Project>
