--- src/tests/tracing/eventpipe/common_legacy/common_legacy.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETCoreApp</TargetFrameworkIdentifier>
-    <OutputType>Library</OutputType>
-    <CLRTestKind>SharedLibrary</CLRTestKind>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <IsTestProject>false</IsTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="IpcTraceTest.legacy.cs" />
-    <Compile Include="IpcUtils.cs" />
-    <Compile Include="Reverse.cs" />
-    <Compile Include="StreamProxy.cs" />
-  </ItemGroup>
-</Project>
