--- src/tests/tracing/eventpipe/gcdump_legacy/gcdump_legacy.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETCoreApp</TargetFrameworkIdentifier>
-    <OutputType>exe</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <UnloadabilityIncompatible>true</UnloadabilityIncompatible>
-    <!-- This test provides no interesting scenarios for GCStress -->
-    <GCStressIncompatible>true</GCStressIncompatible>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildProjectName).cs" />
-    <ProjectReference Include="../common_legacy/common_legacy.csproj" />
-  </ItemGroup>
-</Project>
