+++ src/tests/tracing/eventpipe/providervalidation/providervalidation.cs
@@ -10,7 +10,11 @@ using System.Threading.Tasks;
 using System.Collections.Generic;
 using Microsoft.Diagnostics.Tracing;
 using Tracing.Tests.Common;
+#if (UPDATED_NETCORE_CLIENT == true)
 using Microsoft.Diagnostics.NETCore.Client;
+#else
+using Microsoft.Diagnostics.Tools.RuntimeClient;
+#endif
 
 namespace Tracing.Tests.ProviderValidation
 {
@@ -28,6 +32,7 @@ namespace Tracing.Tests.ProviderValidation
             // This test validates that the rundown events are present
             // and that providers turned on that generate events are being written to
             // the stream.
+#if (UPDATED_NETCORE_CLIENT == true)
             var providers = new List<EventPipeProvider>()
             {
                 new EventPipeProvider("MyEventSource", EventLevel.Verbose),
@@ -35,6 +40,17 @@ namespace Tracing.Tests.ProviderValidation
             };
 
             var ret = IpcTraceTest.RunAndValidateEventCounts(_expectedEventCounts, _eventGeneratingAction, providers, 1024);
+#else
+            var providers = new List<Provider>()
+            {
+                new Provider("MyEventSource"),
+                new Provider("Microsoft-DotNETCore-SampleProfiler")
+            };
+
+            var config = new SessionConfiguration(circularBufferSizeMB: (uint)Math.Pow(2, 10), format: EventPipeSerializationFormat.NetTrace,  providers: providers);
+
+            var ret = IpcTraceTest.RunAndValidateEventCounts(_expectedEventCounts, _eventGeneratingAction, config);
+#endif
             if (ret < 0)
                 return ret;
             else
