--- src/tests/tracing/eventpipe/rundownvalidation_legacy/rundownvalidation_legacy.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworkIdentifier>.NETCoreApp</TargetFrameworkIdentifier>
-    <OutputType>exe</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <JitOptimizationSensitive>true</JitOptimizationSensitive>
-    <!-- The test launches a secondary process and process launch creates
-    an infinite event loop in the SocketAsyncEngine on Linux. Since
-    runincontext loads even framework assemblies into the unloadable
-    context, locals in this loop prevent unloading -->
-    <UnloadabilityIncompatible>true</UnloadabilityIncompatible>
-    <!-- Times out -->
-    <GCStressIncompatible Condition="'$(TargetArchitecture)' == 'arm64' and '$(TargetOS)' == 'OSX'">true</GCStressIncompatible>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="$(MSBuildProjectName).cs" />
-    <ProjectReference Include="../common_legacy/common_legacy.csproj" />
-  </ItemGroup>
-</Project>
